requirements.txt
requests>=2.31.0
python-dateutil>=2.9.0
feedparser>=6.0.11
trafilatura>=1.6.4
lxml>=4.9.0
openai>=1.12.0
PyYAML>=6.0
